+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.3.0] - Unreleased
+
+### Added
+
+- Added Console.size setter
+- Added Console.width setter
+- Added Console.height setter
+
 ## [10.2.2] - 2021-05-19
 
 ### Fixed